django/requirements.txt
Django>=4.2.0
uvicorn
psycopg2-binary
django-environ
django-oauth-toolkit
djangorestframework
channels
websockets
channels_redis
django-cors-headers
pyotp
qrcode
pillow
quote
PyJWT
cryptography
djangorestframework-simplejwt
web3==6.11.1
eth-account==0.9.0
eth-utils==2.3.0
django-prometheus
celery[redis]
django-celery-results
django-redis
bleach